Segregate feature source files and add gulp tasks to process them
