+++ features/2018/designing-for-the-sea-and-stars/designing-for-the-sea-and-stars.js
@@ -1,233 +1,1 @@
-/**
- * DESIGNING FOR THE SEA AND STARS
- *
- * @package uri-modern-home
- */
-
-( function() {
-
-	'use strict';
-
-	window.addEventListener( 'load', initPage, false );
-
-	/**
-	 * Initialize the page
-	 */
-	function initPage() {
-
-		initParallax();
-		initProgressBar();
-
-	}
-
-	/**
-	 * Initialize Parallax
-	 */
-	function initParallax() {
-
-		var frames, top, vh, vw, s, yPos, y, i;
-
-		frames = document.querySelectorAll( '.parallax' );
-
-		window.addEventListener(
-			 'scroll',
-			function() {
-
-			top = this.pageYOffset;
-			vh = this.innerHeight;
-			vw = this.innerWidth;
-
-			for ( i = 0; i < frames.length; i++ ) {
-
-					// Shift the frame offset so that the bottom of the viewport is now 0
-					// This makes it so that frames don't move until they're scrolled into view
-					y = frames[i].getBoundingClientRect().top - vh;
-
-					// Adjust the speed to be relative to the viewport width
-					s = frames[i].getAttribute( 'data-speed' ) * vw / 1000;
-
-					// Set the y translation of the frame (only generates numbers <=0)
-					yPos = Math.min( 0, ( y * s / 100 ) );
-					frames[i].setAttribute( 'style', 'transform: translate3d(0px, ' + yPos + 'px, 0px)' );
-			}
-
-		}
-			);
-
-	}
-
-	/**
-	 * Initialize the progress bar
-	 */
-	function initProgressBar() {
-
-		// Define a bunch of stuff we can pass around inside one object
-		var data = {
-			vh: window.innerHeight,
-			vw: window.innerWidth,
-			lastPos: window.pageYOffset,
-			bar: document.getElementById( 'progress' ),
-			indicator: document.getElementById( 'indicator' ),
-			scale: getProgressScale(),
-			sky: document.getElementById( 'keyframe-backdrop-sky' ),
-			story: document.getElementById( 'story' )
-		};
-
-		// Get some measurements
-		data.pos = data.story.getBoundingClientRect().top;
-		data.storyHeight = data.story.offsetHeight;
-
-		// Add a hook now that we have js
-		data.bar.classList.add( 'has-js' );
-
-		// Initialize the scale and update the indicator
-		setProgressScale( data );
-		updateProgress( data );
-
-		// Update indicator on scroll
-		window.addEventListener( 'scroll', updateProgress.bind( null, data ), false );
-
-		// Update measurements, scale, and indicator on window resize
-		window.addEventListener(
-			 'resize',
-			function() {
-
-			data.vh = window.innerHeight;
-			data.vw = window.innerWidth;
-
-			data.storyHeight = data.story.offsetHeight;
-
-			setProgressScale( data );
-			updateProgress( data );
-
-		},
-			false
-			);
-
-	}
-
-	/**
-	 * Get the progress scale labels and depths
-	 */
-	function getProgressScale() {
-
-		var scale, i;
-
-		scale = {
-			ticks: document.getElementById( 'progress' ).querySelectorAll( '.progress-label' ),
-			depths: []
-		};
-
-		for ( i = 0; i < scale.ticks.length; i++ ) {
-			scale.depths.push( scale.ticks[i].getAttribute( 'data-depth' ) );
-		}
-
-		return scale;
-
-	}
-
-	/**
-	 * Set/update the progress scale
-	 *
-	 * @param data obj The data initialized in initProgressBar().
-	 */
-	function setProgressScale( data ) {
-
-		var n, i, bottom;
-
-		n = data.scale.ticks.length;
-
-		// Determine how far down the page (in percent) the waterline is
-		data.scale.waterline = data.sky.offsetHeight / data.storyHeight * 100;
-
-		// Define the offset (in percent) for the last tickmark
-		bottom = 98;
-
-		// Set the progress bar offset margin
-		data.bar.style.marginLeft = 'calc( -50vw + (' + getScrollbarWidth() + 'px / 2 ) )';
-
-		// Set the background colors of the progress bar
-		data.bar.style.background = 'linear-gradient(to bottom, rgba(0,0,0,.4) calc( ' + data.scale.waterline + '% + 5px ), rgba(18,85,146,.4) calc( ' + data.scale.waterline + '% + 5px ), rgba(0,20,40,.6) calc( 98% + 5px ), rgba(120,96,38,.5) calc( 98% + 5px ))';
-
-		// Set the offsets for the first and last tickmark (in percent)
-		data.scale.ticks[0].style.top = data.scale.waterline + '%';
-		data.scale.ticks[n - 1].style.top = bottom + '%';
-
-		// Loop through the other tickmarks and calculate/set their offsets (in percent)
-		for ( i = 1; i < n - 1; i++ ) {
-			data.scale.ticks[i].style.top = data.scale.waterline + ( data.scale.depths[i] / data.scale.depths[n - 1] ) * ( bottom - data.scale.waterline ) + '%';
-		}
-
-	}
-
-	/**
-	 * Update the progress bar
-	 *
-	 * @param data obj The data initialized in initProgressBar().
-	 */
-	function updateProgress( data ) {
-
-		var yPos, prog;
-
-		yPos = window.pageYOffset;
-
-		// Calculate the progress as a percent between 0-97
-		prog = Math.min( 97, Math.max( 0, yPos / data.storyHeight * 100 ) );
-		data.indicator.style.top = prog + '%';
-
-		// Swap out the rocket for the submarine at the waterline
-		prog > data.scale.waterline - 1 ? data.indicator.classList.add( 'titan' ) : data.indicator.classList.remove( 'titan' );
-
-		// Swap out the rocket flame for the parachute based on scroll direction
-		yPos < data.lastPos ? data.indicator.classList.add( 'flame' ) : data.indicator.classList.remove( 'flame' );
-
-		// Store the scroll position for later reference in determining scroll direction
-		data.lastPos = yPos;
-
-		// If the page scroll is past the initial story top, make the progress bar sticky.
-		// And, if the end of the story scrolls into view, move the progress bar along with it.
-		if ( yPos > data.pos ) {
-			data.bar.classList.add( 'sticky' );
-			data.storyHeight + data.pos < yPos + data.vh ? data.bar.classList.add( 'end' ) : data.bar.classList.remove( 'end' );
-		} else {
-			data.bar.classList.remove( 'sticky' );
-		}
-
-	}
-
-	/*
-	 * Get scrollbar width
-	 *
-	 * @see http://jsfiddle.net/UU9kg/17/
-	 */
-	function getScrollbarWidth() {
-
-		var outer, widthNoScroll, inner, widthWithScroll;
-
-		outer = document.createElement( 'div' );
-		outer.style.visibility = 'hidden';
-		outer.style.width = '100px';
-		outer.style.msOverflowStyle = 'scrollbar'; // Needed for WinJS apps
-
-		document.body.appendChild( outer );
-
-		widthNoScroll = outer.offsetWidth;
-
-		// Force scrollbars
-		outer.style.overflow = 'scroll';
-
-		// Add innerdiv
-		inner = document.createElement( 'div' );
-		inner.style.width = '100%';
-		outer.appendChild( inner );
-
-		widthWithScroll = inner.offsetWidth;
-
-		// Remove divs
-		outer.parentNode.removeChild( outer );
-
-		return widthNoScroll - widthWithScroll;
-
-	}
-
-})();
+/* built */!function(){"use strict";function o(t){var e,s,i,a,n,r;for(e=t.scale.ticks.length,t.scale.waterline=t.sky.offsetHeight/t.storyHeight*100,98,t.bar.style.marginLeft="calc( -50vw + ("+((i=document.createElement("div")).style.visibility="hidden",i.style.width="100px",i.style.msOverflowStyle="scrollbar",document.body.appendChild(i),a=i.offsetWidth,i.style.overflow="scroll",(n=document.createElement("div")).style.width="100%",i.appendChild(n),r=n.offsetWidth,i.parentNode.removeChild(i),a-r)+"px / 2 ) )",t.bar.style.background="linear-gradient(to bottom, rgba(0,0,0,.4) calc( "+t.scale.waterline+"% + 5px ), rgba(18,85,146,.4) calc( "+t.scale.waterline+"% + 5px ), rgba(0,20,40,.6) calc( 98% + 5px ), rgba(120,96,38,.5) calc( 98% + 5px ))",t.scale.ticks[0].style.top=t.scale.waterline+"%",t.scale.ticks[e-1].style.top="98%",s=1;s<e-1;s++)t.scale.ticks[s].style.top=t.scale.waterline+t.scale.depths[s]/t.scale.depths[e-1]*(98-t.scale.waterline)+"%"}function d(t){var e,s;e=window.pageYOffset,s=Math.min(97,Math.max(0,e/t.storyHeight*100)),t.indicator.style.top=s+"%",s>t.scale.waterline-1?t.indicator.classList.add("titan"):t.indicator.classList.remove("titan"),e<t.lastPos?t.indicator.classList.add("flame"):t.indicator.classList.remove("flame"),(t.lastPos=e)>t.pos?(t.bar.classList.add("sticky"),t.storyHeight+t.pos<e+t.vh?t.bar.classList.add("end"):t.bar.classList.remove("end")):t.bar.classList.remove("sticky")}window.addEventListener("load",function(){e=document.querySelectorAll(".parallax"),window.addEventListener("scroll",function(){for(this.pageYOffset,s=this.innerHeight,i=this.innerWidth,l=0;l<e.length;l++)r=e[l].getBoundingClientRect().top-s,a=e[l].getAttribute("data-speed")*i/1e3,n=Math.min(0,r*a/100),e[l].setAttribute("style","transform: translate3d(0px, "+n+"px, 0px)")}),t={vh:window.innerHeight,vw:window.innerWidth,lastPos:window.pageYOffset,bar:document.getElementById("progress"),indicator:document.getElementById("indicator"),scale:function(){var t,e;for(t={ticks:document.getElementById("progress").querySelectorAll(".progress-label"),depths:[]},e=0;e<t.ticks.length;e++)t.depths.push(t.ticks[e].getAttribute("data-depth"));return t}(),sky:document.getElementById("keyframe-backdrop-sky"),story:document.getElementById("story")},t.pos=t.story.getBoundingClientRect().top,t.storyHeight=t.story.offsetHeight,t.bar.classList.add("has-js"),o(t),d(t),window.addEventListener("scroll",d.bind(null,t),!1),window.addEventListener("resize",function(){t.vh=window.innerHeight,t.vw=window.innerWidth,t.storyHeight=t.story.offsetHeight,o(t),d(t)},!1);var t;var e,s,i,a,n,r,l},!1)}();
