+++ features/2018/designing-for-the-sea-and-stars/src/designing-for-the-sea-and-stars-homepage.scss
@@ -1,0 +1,47 @@
+/* ==============================
+   DESIGNING FOR THE SEA AND STARS HOMEPAGE HERO
+   ============================== */
+
+@import url("https://fonts.googleapis.com/css?family=Julius+Sans+One|Economica");
+
+#feature-hero .still {
+	background-position: center bottom;
+}
+
+#feature-hero h1 {
+	max-width: 600px;
+	margin: 0 auto 2rem;
+	font-family: 'Julius Sans One', 'Lato', cursive, sans-serif;
+	font-size: 4rem;
+	color: #fff;
+}
+
+#feature-hero h1 span {
+	display: block;
+	font-size: 1.5rem;
+	margin-bottom: 1rem;
+}
+
+@media screen and (min-width: 360px) {
+
+	#feature-hero h1 {
+		font-size: 5rem;
+	}
+
+	#feature-hero h1 span {
+		font-size: 2rem;
+	}
+
+}
+
+@media screen and (min-width: 420px) {
+
+	#feature-hero h1 {
+		font-size: 7rem;
+	}
+
+	#feature-hero h1 span {
+		font-size: 3rem;
+	}
+
+}
