+++ features/2018/the-internationalists/src/the-internationalists.scss
@@ -1,0 +1,366 @@
+/* ==============================
+   THE INTERNATIONALISTS
+   ============================== */
+
+@import url("https://fonts.googleapis.com/css?family=Tulpen+One|Raleway:200");
+
+.globalnav {
+	display: none;
+}
+
+#hero .cl-hero {
+	background: #000;
+	margin-bottom: 0;
+}
+
+#hero h1 {
+	font-family: "Tulpen One", "Hind", "Lato", sans-serif;
+	letter-spacing: 1.2vw;
+	font-size: 2rem;
+	max-width: 100%;
+	text-align: center;
+	text-transform: lowercase;
+	margin-bottom: 0;
+	color: #FFC97B;
+}
+
+@media screen and (min-width: 600px) {
+	#hero h1 {
+		font-size: 2.5rem;
+	}
+}
+
+@media screen and (min-width: 800px) {
+	#hero h1 {
+		font-size: 3.2rem;
+	}
+}
+
+@supports ( (-webkit-background-clip: initial) and (-webkit-text-fill-color: initial) ) {
+
+	#hero h1 {
+		color: transparent;
+		background: linear-gradient(to bottom, #C44644, #EB9E69);
+		-webkit-background-clip: text;
+		-webkit-text-fill-color: transparent;
+		text-shadow: none;
+	}
+}
+
+#hero .cl-hero .still {
+	background-size: 100% auto;
+}
+
+.artboard {
+	margin-bottom: 3rem;
+	position: relative;
+	left: -25px;
+	top: 0;
+}
+
+.artboard.section-break {
+	margin-bottom: 0;
+}
+
+.artboard img {
+	width: 100%;
+	height: auto;
+	padding: 0;
+	position: relative;
+}
+
+.artboard.reveal.has-js {
+	visibility: hidden;
+	top: 2rem;
+	opacity: 0;
+	transition: top 1s ease-in-out, opacity .5s ease-in-out;
+}
+
+.artboard.reveal.has-js img {
+	filter: brightness(10);
+	transition: filter .8s ease-in-out;
+}
+
+.artboard.reveal.has-js.visible {
+	visibility: visible;
+	top: 0;
+	opacity: 1;
+}
+
+.artboard.reveal.has-js.visible img {
+	filter: brightness(1);
+}
+
+.caption {
+	font-size: .9rem;
+	color: rgba(0,0,0,.5);
+	padding: 1rem 25px 0;
+	max-width: 360px;
+}
+
+.section-break .caption {
+	position: absolute;
+	bottom: -2rem;
+	left: 25px;
+	padding: 0;
+	z-index: 10;
+}
+
+@media screen and (max-width: 1000px) {
+	.artboard {
+		width: 100vw;
+	}
+}
+
+@media screen and (min-width: 1000px) {
+	.artboard {
+		margin-bottom: 4rem;
+		left: 0;
+	}
+	.caption {
+		padding: 1rem 0;
+	}
+}
+
+.content-canvas {
+	padding-top: 5rem;
+	padding-bottom: 3rem;
+}
+
+.content-canvas h1 {
+	font-family: "Tulpen One", "Hind", "Lato", sans-serif;
+	font-size: 3rem;
+	margin-bottom: 3rem;
+	letter-spacing: .5rem;
+	text-transform: lowercase;
+}
+
+.content-body {
+	max-width: 700px;
+	margin-left: auto;
+	margin-right: auto;
+	margin-bottom: 4rem;
+}
+
+.section-lead {
+	background: linear-gradient(#000, #3a2b24);
+	color: #ccc;
+	padding-top: 0;
+}
+
+.content-deck {
+	font-size: 1.4rem;
+	font-family: "Raleway", "Hind", sans-serif;
+	margin-bottom: 4rem;
+	padding-bottom: 2rem;
+	color: #c19883;
+}
+
+.content-deck:after {
+	display: block;
+	content: '';
+	width: 100%;
+	position: relative;
+	top: 2rem;
+	height: 1px;
+	background: linear-gradient(45deg, #7c5928, #913b34);
+}
+
+.content-deck p {
+	width: 100%;
+	max-width: 600px;
+	margin: 0 auto 2rem;
+	text-align: center;
+}
+
+.section-raising-the-bar {
+	background: linear-gradient(#f4a974, #f0c17f);
+	color: #641e07;
+}
+
+.section-raising-the-bar h1 {
+	color: #ab4506;
+}
+
+.section-interwoven-cultures {
+	background: linear-gradient(#fff7f7, #f0d6dd);
+	color: #55022c;
+}
+
+.section-interwoven-cultures h1 {
+	color: #8e5863;
+}
+
+.section-visionary-programs {
+	background: linear-gradient(#3d1b01, #000);
+	color: #c99a94;
+}
+
+.section-visionary-programs h1 {
+	color: #c46a54;
+}
+
+.section-visionary-programs a {
+	color: #ec6863;
+}
+
+.section-visionary-programs p a {
+	text-decoration: none;
+}
+
+.section-visionary-programs p a:hover {
+	color: #ec6863;
+	text-decoration: underline;
+}
+
+.section-uniquely-prepared {
+	background: linear-gradient(#efffff, rgba(239,255,255,0) );
+}
+
+.intl-programs {
+	display: flex;
+	flex-wrap: wrap;
+	justify-content: left;
+	text-align: center;
+	font-family: "Hind", sans-serif;
+	font-size: .9rem;
+	margin-top: 1.5rem;
+	margin-bottom: 4rem;
+}
+
+.intl-programs>div {
+	flex: 1;
+	box-sizing: border-box;
+	min-width: 150px;
+	max-width: 200px;
+	padding: 0 1.5rem;
+}
+
+.intl-programs a {
+	color: #aa7161;
+	box-sizing: border-box;
+	display: block;
+	border: 1px solid transparent;
+	padding: 1rem .5rem;
+	border-radius: 5px;
+	transition: all .1s ease-in-out;
+}
+
+.intl-programs a:hover,
+.intl-programs a:focus {
+	border: 1px solid rgba(255,255,255,.2);
+}
+
+.intl-programs span {
+	display: block;
+}
+
+.intl-programs img {
+	max-width: 100%;
+	min-width: initial;
+	padding: 0;
+	width: 64px;
+	margin: 0 auto 1.5rem;
+}
+
+.section-post-script {
+	font-family: "Hind", Arial, sans-serif;
+	font-size: 1rem;
+	color: #555;
+	position: relative;
+	padding: 2rem 0 0;
+}
+
+.section-post-script:before {
+	display: block;
+	content: '';
+	width: 100px;
+	height: 1px;
+	background: #ddd;
+	position: absolute;
+	top: 0;
+}
+
+.section-post-script p {
+	max-width: 400px;
+}
+
+.primary-action-promo {
+	display: block;
+	max-width: 600px;
+	margin: 0 auto 6rem;
+	padding: 2rem;
+	box-sizing: border-box;
+	background: #fff;
+	border-radius: .2rem;
+	position: relative;
+	top: 0;
+	box-shadow: 0 .1rem 3rem rgba(0,0,0,.1);
+	transition: all .2s ease-in-out;
+}
+
+.primary-action-promo:after {
+	display: block;
+	content: '';
+	clear: both;
+}
+
+.primary-action-promo:hover {
+	top: -2px;
+	box-shadow: 0 .1rem 1rem rgba(0,0,0,.2);
+	transition: all .1s ease-in-out;
+}
+
+.primary-action-promo .promo-header {
+	display: block;
+	font-family: "Lato","Hind",Arial,sans-serif;
+	font-weight: 700;
+	font-size: 1.875rem;
+	margin-bottom: 2rem;
+	line-height: 1.2;
+	color: #b5d3e0;
+}
+
+.primary-action-promo .promo-text {
+	display: block;
+	font-family: "Hind", Arial, sans-serif;
+	font-size: 1.2rem;
+	line-height: 1.5;
+	color: #333;
+	max-width: 400px;
+}
+
+.primary-action-promo .promo-link {
+	color: inherit;
+	font-weight: 700;
+	white-space: nowrap;
+}
+
+.primary-action-promo .promo-stamp {
+	position: absolute;
+	top: 30px;
+	right: 40px;
+	opacity: .3;
+	transform: rotate(30deg);
+	width: 0;
+	height: 100px;
+	overflow: hidden;
+	padding-left: 100px;
+	background-repeat: no-repeat;
+	background-position: center center;
+	background-size: 100% 100%;
+}
+
+@media screen and (min-width: 600px) {
+
+	.primary-action-promo .promo-header {
+		float: left;
+		max-width: 30%;
+	}
+
+	.primary-action-promo .promo-text {
+		float: right;
+		max-width: 60%;
+	}
+
+} /* end min-width:600px */
