+++ features/2018/the-internationalists/the-internationalists-homepage.css
@@ -1,61 +1,1 @@
-/* ==============================
-   THE INTERNATIONALISTS HOMEPAGE HERO
-   ============================== */
-
-@import url("https://fonts.googleapis.com/css?family=Tulpen+One");
-
-#feature-hero .block {
-	background: none;
-}
-
-#feature-hero h1 {
-	font-family: "Tulpen One", "Hind", "Lato", sans-serif;
-	letter-spacing: 1.2vw;
-	font-size: 2.9rem;
-	max-width: 100%;
-	text-align: center;
-	text-transform: lowercase;
-	padding-left: 0;
-	padding-right: 0;
-	color: #FFC97B;
-}
-
-#feature-hero .cl-hero .still {
-	background-size: auto 165%;
-	background-position: 48% 53%;
-}
-
-@media screen and (min-width: 600px) {
-	#feature-hero h1 {
-		font-size: 3.4rem;
-	}
-	#feature-hero .cl-hero .still {
-		background-size: cover;
-		background-position: center center;
-	}
-}
-
-@media screen and (min-width: 800px) {
-	#feature-hero h1 {
-		font-size: 4rem;
-	}
-}
-
-@supports ( (-webkit-background-clip: initial) and (-webkit-text-fill-color: initial) ) {
-
-	#feature-hero h1 {
-		color: transparent;
-		background: linear-gradient(to bottom, #C44644, #EB9E69);
-		-webkit-background-clip: text;
-		-webkit-text-fill-color: transparent;
-		text-shadow: none;
-	}
-}
-
-#feature-hero .cl-hero {
-	background: #000;
-}
-
-#feature-hero .cl-hero .cl-button {
-	max-width: 360px;
-}
+/* built */@import url("https://fonts.googleapis.com/css?family=Tulpen+One");#feature-hero .block{background:none}#feature-hero h1{font-family:"Tulpen One", "Hind", "Lato", sans-serif;letter-spacing:1.2vw;font-size:2.9rem;max-width:100%;text-align:center;text-transform:lowercase;padding-left:0;padding-right:0;color:#FFC97B}#feature-hero .cl-hero .still{background-size:auto 165%;background-position:48% 53%}@media screen and (min-width: 600px){#feature-hero h1{font-size:3.4rem}#feature-hero .cl-hero .still{background-size:cover;background-position:center center}}@media screen and (min-width: 800px){#feature-hero h1{font-size:4rem}}@supports (-webkit-background-clip: initial) and (-webkit-text-fill-color: initial){#feature-hero h1{color:transparent;background:-webkit-gradient(linear, left top, left bottom, from(#C44644), to(#EB9E69));background:linear-gradient(to bottom, #C44644, #EB9E69);-webkit-background-clip:text;-webkit-text-fill-color:transparent;text-shadow:none}}#feature-hero .cl-hero{background:#000}#feature-hero .cl-hero .cl-button{max-width:360px}
