+++ features/2018/the-internationalists/the-internationalists.js
@@ -1,52 +1,1 @@
-/**
- * THE INTERNATIONALISTS
- *
- * @package uri-modern-home
- */
-
-( function() {
-
-	'use strict';
-
-	var els, l;
-
-	window.addEventListener(
-		 'load',
-		function() {
-
-		var i;
-
-		els = document.querySelectorAll( '.artboard.reveal' );
-		l = els.length;
-
-		for ( i = 0; i < l; i++ ) {
-				els[i].classList.add( 'has-js' );
-		}
-
-		showImages;
-
-	}
-		);
-
-	window.addEventListener( 'scroll', showImages, false );
-	window.addEventListener( 'resize', showImages, false );
-
-	function showImages() {
-
-		var vh, offset, i;
-
-		vh = window.innerHeight;
-
-		for ( i = 0; i < l; i++ ) {
-
-			offset = els[i].getBoundingClientRect().top;
-
-			if ( vh * .8 > offset ) {
-				els[i].classList.add( 'visible' );
-			}
-
-		}
-
-	}
-
-})();
+/* built */!function(){"use strict";var t,i;function e(){var e,n;for(e=window.innerHeight,n=0;n<i;n++)t[n].getBoundingClientRect().top<.8*e&&t[n].classList.add("visible")}window.addEventListener("load",function(){var e;for(t=document.querySelectorAll(".artboard.reveal"),i=t.length,e=0;e<i;e++)t[e].classList.add("has-js")}),window.addEventListener("scroll",e,!1),window.addEventListener("resize",e,!1)}();
