+++ features/2018/the-philosophy-of-comedy/src/the-philosophy-of-comedy.scss
@@ -1,0 +1,287 @@
+/* ==============================
+   PHILOSOPHY OF COMEDY
+   ============================== */
+
+@import url('https://fonts.googleapis.com/css?family=Josefin+Slab:400,700');
+
+.globalnav {
+	display: none;
+}
+
+h1.title {
+	font-family: "Josefin Slab", "Hind", serif;
+	font-size: 16.5vw;
+	font-weight: 700;
+	color: #fff;
+	text-shadow: 3px 3px 0 #ff6872; /* #376cdd */
+}
+
+h1.title span {
+	color: #fff;
+	background: #ff6872; /* #ffff80 */
+	text-shadow: none;
+}
+
+@media screen and (min-width: 700px ) {
+	h1.title {
+		font-size: 8vw;
+	}
+}
+
+h1:not(.title) {
+	font-family: "Josefin Slab", "Lato", "Hind", sans-serif;
+	font-size: 3rem;
+	font-weight: 400;
+}
+
+h1,
+h2 {
+	max-width: 100%;
+}
+
+h2 {
+	display: inline-block;
+	color: #fff;
+	background: #ff6872;
+	padding: .5rem;
+	box-shadow: 5px 5px 0 #555;
+}
+
+.copyblock {
+	max-width: 700px;
+	margin-left: auto;
+	margin-right: auto;
+	margin-bottom: 4rem;
+}
+
+.copyblock p {
+	line-height: 1.8;
+	-webkit-font-smoothing: auto;
+}
+
+.headerblock {
+	border-bottom: 1px solid #ddd;
+	margin-bottom: 3rem;
+	text-align: center;
+}
+
+.artboard {
+	margin-bottom: 2rem;
+	position: relative;
+	left: -25px;
+	width: 100vw;
+}
+
+.artboard img {
+	width: 100%;
+	height: auto;
+	padding: 0;
+}
+
+.caption {
+	font-size: .9rem;
+	color: #555;
+	padding: 1rem 25px 0;
+	max-width: 360px;
+}
+
+.caption:after {
+	display: block;
+	content: '';
+	width: 50px;
+	height: 1px;
+	background: #ff6872;
+	margin-top: 1rem;
+}
+
+.caption p {
+	line-height: 125%;
+	margin-bottom: 1rem;
+}
+
+.caption a {
+	text-decoration: none;
+}
+
+.caption a:hover,
+.caption a:focus {
+	text-decoration: underline;
+}
+
+.credit {
+	color: #999;
+	font-family: "Hind", serif;
+	font-weight: 300;
+	font-size: .75rem;
+	margin-bottom: 1rem;
+}
+
+@media screen and (min-width: 1000px) {
+	.artboard {
+		margin-bottom: 3rem;
+		left: 0;
+		width: 100%;
+	}
+	.caption {
+		padding: 1rem 0;
+	}
+}
+
+#hero div:nth-child(1) {
+	width: 100%;
+}
+
+#hero div:nth-child(2) {
+	position: absolute;
+	left: 11vw;
+	top: 45%;
+	z-index: 10;
+}
+
+#hero img {
+	width: 200%;
+	max-width: initial;
+	height: auto;
+	margin-left: -50%;
+	padding: 0;
+	display: block;
+}
+
+@keyframes feature-title-animate {
+	0% { transform: rotate(0); }
+	20% { transform: rotate(26deg); }
+	30% { transform: rotate(18deg); }
+	40% { transform: rotate(25deg); }
+	50% { transform: rotate(21deg); }
+	60% { transform: rotate(24deg); }
+	70% { transform: rotate(22deg); }
+	80% { transform: rotate(23deg); }
+	90% { transform: rotate(22.5deg); }
+	100% { transform: rotate(23deg); }
+}
+
+@media screen and (min-width: 700px) {
+	#hero div:nth-child(2) {
+		left: 58%;
+		top: 40%;
+		transform: rotate(0);
+		transform-origin: 0 0;
+		animation-name: feature-title-animate;
+		animation-delay: 1.5s;
+		animation-duration: 3s;
+		animation-timing-function: ease-in-out;
+		animation-fill-mode: forwards;
+	}
+	#hero img {
+		width: 100%;
+		margin-left: 0;
+	}
+}
+
+#deck {
+	background: #ff6872;
+	padding: 12vw 1rem 5vw 0;
+	margin-bottom: 2rem;
+	box-sizing: border-box;
+	transform: rotate(-2deg);
+	position: relative !important;
+	width: 140vw !important;
+	left: 70% !important;
+	margin-left: -70vw !important;
+	top: -3vw;
+}
+
+#deck>div {
+	font-family: "Josefin Slab", "Lato", "Hind", sans-serif;
+	font-size: 1.5rem;
+	width: 75vw;
+	max-width: 500px;
+	border-left: 10px solid #fff;
+	padding-left: 2rem;
+	color: #fff;
+	transform: rotate(2deg);
+	margin-left: 2rem;
+}
+
+@media screen and (min-width: 700px) { /* Must be 700 */
+	#deck {
+		padding: 4rem 0 1rem;
+	}
+	#deck>div {
+		width: initial;
+		font-size: 1.8rem;
+		margin-left: calc(55vw - 350px);
+	}
+}
+
+.cl-tiles.halves.compact>div:nth-child(1) {
+	margin-bottom: .5rem;
+}
+
+@media screen and (min-width: 800px) {
+	.cl-tiles.halves.compact>div:nth-child(1) {
+		margin-bottom: 0;
+	}
+}
+
+.two-to-one {
+	display: block;
+}
+
+.two-to-one div:nth-child(1) {
+	width: 100%;
+	margin-bottom: .5rem;
+}
+
+.two-to-one div:nth-child(2) {
+	width: 100%;
+}
+
+.two-to-one img {
+	display: block;
+	padding: 0;
+}
+
+@media screen and (min-width: 600px) {
+
+	.two-to-one {
+		display: flex;
+		justify-content: space-between;
+	}
+
+	.two-to-one div:nth-child(1) {
+		width: 30.45%;
+	}
+
+	.two-to-one div:nth-child(2) {
+		width: 68.5%;
+	}
+
+}
+
+#coffee-stain {
+	position: relative;
+}
+
+#coffee-stain img {
+	position: absolute;
+	top: -10rem;
+	left: -35%;
+	width: 75%;
+	min-width: 500px;
+	height: auto;
+	padding: 0;
+}
+
+#chickenfeet {
+	top: -20vw;
+	margin-bottom: -40vw;
+	z-index: 10;
+	pointer-events: none;
+}
+
+#chickenfeet img {
+	width: 100%;
+	height: auto;
+	padding: 0;
+}
