+++ features/2018/uri-a-users-guide/uri-a-users-guide.css
@@ -1,89 +1,1 @@
-/* ==============================
-   URI: A USER'S GUIDE 2018
-   ============================== */
-
-@import url(https://fonts.googleapis.com/css?family=Gochi+Hand);
-
-.content-canvas {
-	padding-top: 5rem;
-	padding-bottom: 3rem;
-}
-
-.content-body {
-	max-width: 700px;
-	margin-left: auto;
-	margin-right: auto;
-	margin-bottom: 4rem;
-}
-
-.content-body h1 {
-	font-family: "Gochi Hand", "Lato", "Hind", sans-serif;
-	font-size: 3rem;
-}
-
-.content-body a:not(.cl-card) {
-	text-decoration: underline;
-}
-
-.cl-card {
-	z-index: 10;
-}
-
-.section-header {
-	padding: 20% 0 18%;
-	background: #f8f8f8;
-}
-
-.section-header h1 {
-	font-size: 4rem;
-	color: #002147;
-	font-weight: 900;
-}
-
-.section-make-yourself-at-home {
-	background: #1ba5ff;
-	color: #fff;
-}
-
-
-.section-find-your-people {
-	background: #153e50;
-	color: #fff;
-}
-
-.section-play {
-	background: #ffff7a;
-}
-
-.section-eat-really-well {
-	background: #2b57dd;
-	color: #fff;
-}
-
-.section-catch-a-ride {
-	background: #a5e2f0;
-}
-
-.section-take-in-a-show {
-	background: #fff6af;
-}
-
-.section-cheer-on-the-rams {
-	background: #3b64bd;
-	color: #fff;
-}
-
-.section-make-yourself-at-home a,
-.section-eat-really-well a,
-.section-cheer-on-the-rams a,
-.section-find-your-people a,
-.section-cheer-on-the-rams a {
-	color: #fffc85;
-}
-
-.section-make-yourself-at-home h1,
-.section-eat-really-well h1,
-.section-find-your-people h1,
-.section-cheer-on-the-rams h1 {
-	color: #fff;
-}
+/* built */@import url(https://fonts.googleapis.com/css?family=Gochi+Hand);.content-canvas{padding-top:5rem;padding-bottom:3rem}.content-body{max-width:700px;margin-left:auto;margin-right:auto;margin-bottom:4rem}.content-body h1{font-family:"Gochi Hand", "Lato", "Hind", sans-serif;font-size:3rem}.content-body a:not(.cl-card){text-decoration:underline}.cl-card{z-index:10}.section-header{padding:20% 0 18%;background:#f8f8f8}.section-header h1{font-size:4rem;color:#002147;font-weight:900}.section-make-yourself-at-home{background:#1ba5ff;color:#fff}.section-find-your-people{background:#153e50;color:#fff}.section-play{background:#ffff7a}.section-eat-really-well{background:#2b57dd;color:#fff}.section-catch-a-ride{background:#a5e2f0}.section-take-in-a-show{background:#fff6af}.section-cheer-on-the-rams{background:#3b64bd;color:#fff}.section-make-yourself-at-home a,.section-eat-really-well a,.section-cheer-on-the-rams a,.section-find-your-people a,.section-cheer-on-the-rams a{color:#fffc85}.section-make-yourself-at-home h1,.section-eat-really-well h1,.section-find-your-people h1,.section-cheer-on-the-rams h1{color:#fff}
