+++ gulpfile.js
@@ -32,6 +32,8 @@ var sass = require('gulp-sass');
 var sourcemaps = require('gulp-sourcemaps');
 var autoprefixer = require('autoprefixer');
 var postcss = require('gulp-postcss');
+var rename = require('gulp-rename');
+var path = require('path');
 var header = require('gulp-header');
 var shell = require('gulp-shell');
 
@@ -66,7 +68,7 @@ function scripts(done) {
  // console.log('scripts ran');
 }
 
-// CSS concat, auto-prefix and minify
+// Theme CSS concat, auto-prefix and minify
 gulp.task('styles', styles);
 
 function styles(done) {
@@ -82,6 +84,43 @@ function styles(done) {
 
   done();
   //console.log('styles ran');
+}
+
+// Features CSS
+gulp.task('featuresCSS', featuresCSS);
+
+function featuresCSS(done) {
+
+	gulp.src('./features/**/src/*.scss')
+		.pipe(sass(sassOptions).on('error', sass.logError))
+		.pipe(rename(function(file) {
+			file.dirname = path.dirname(file.dirname);
+			return file;
+		}))
+        .pipe(postcss([ autoprefixer() ]))
+		.pipe(header('/* built */'))
+		.pipe(gulp.dest('./features/'));
+	
+	done();
+	//console.log('features css ran');
+}
+
+// Features JS
+gulp.task('featuresJS', featuresJS);
+
+function featuresJS(done) {
+
+	gulp.src('./features/**/src/*.js')
+		.pipe(rename(function(file) {
+			file.dirname = path.dirname(file.dirname);
+			return file;
+		}))
+    	.pipe(uglify())
+		.pipe(header('/* built */'))
+    	.pipe(gulp.dest('./features/'));
+
+	done();
+	//console.log('features js ran');
 }
 
 // minify new images
@@ -116,8 +155,14 @@ function watcher(done) {
 	// watch for JS changes
 	gulp.watch('./src/js/*.js', scripts);
 
-	// watch for CSS changes
+	// watch for Theme CSS changes
 	gulp.watch('./src/sass/**/*', styles);
+	
+	// watch for Features CSS changes
+	gulp.watch('./features/**/src/*.scss', featuresCSS);
+	
+	// watch for Features JS changes
+	gulp.watch('./features/**/src/*.js', featuresJS);
 
 	// watch for image changes
 	gulp.watch('./src/images/**/*', images);
@@ -129,7 +174,7 @@ function watcher(done) {
 }
 
 gulp.task( 'default',
-	gulp.parallel('images', 'scripts', 'styles', 'sniffs', 'watcher', function(done){
+	gulp.parallel('images', 'scripts', 'styles', 'featuresCSS', 'featuresJS', 'sniffs', 'watcher', function(done){
 		done();
 	})
 );
