+++ package.json
@@ -29,6 +29,7 @@
     "gulp-jscs": "^4.1.0",
     "gulp-jshint": ">=2.0.4",
     "gulp-postcss": "^8.0.0",
+    "gulp-rename": "^1.4.0",
     "gulp-sass": ">=3.0.0",
     "gulp-shell": "^0.6.5",
     "gulp-sourcemaps": "^2.6.4",
